tg_bot/requirements.txt
python-telegram-bot==20.0
pika==1.3.2
pydantic==1.10.12
mplfinance==0.12.10b0
matplotlib==3.7.2
pandas==2.0.3
numpy==1.26.4
aio-pika>=9.4.1